Dev urvish

+++ app/Http/Controllers/ProfileController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Quiz;
 use App\Models\SubscribeCourse;
 use App\Models\User;
 use Illuminate\Http\Request;
@@ -14,7 +15,7 @@ class ProfileController extends BaseController
     {
         try {
             $id = auth()->user()->id;
-            $user = User::with('course')->where('id', $id)->first();
+            $user = User::where('is_deleted', 0)->with('course')->where('id', $id)->first();
             return $this->sendSuccess($user, "User Data");
         } catch (\Throwable $th) {
             return $this->sendError("Internal Server Error", 500);
@@ -34,20 +35,25 @@ class ProfileController extends BaseController
                 'country' => 'required'
             ]);
 
+            // check email already exist
+            if (User::where('email', $request->email)->where('id', '<>', auth()->user()->id)->first()) {
+                return $this->sendError("Email already exists", 409);
+            }
+
             // validation error
             if ($validation->fails()) {
                 return $this->sendError("Validation Error", 403);
             }
 
-            if ($request->user()->tokenCan('auth-edit-profile')){
+            if ($request->user()->tokenCan('auth-edit-profile')) {
                 // store details
                 $user = User::find(auth()->user()->id);
-                
+
                 if ($request->hasFile('avatar')) {
                     $avatar_url = $this->upload('avatar', 'avatar');
                     $user->avatar = $avatar_url;
                 }
-                
+
                 $user->name = $request->name;
                 $user->email = $request->email;
                 $user->mobile = $request->mobile;
@@ -56,34 +62,36 @@ class ProfileController extends BaseController
                 $user->country = $request->country;
                 $user->save();
 
-                $courseArr = explode(",",$request->courses);
-                $subCourse = SubscribeCourse::where('user_id', $user->id)->get();
+                if ($request->courses) {
+                    $courseArr = explode(",", $request->courses);
+                    $subCourse = SubscribeCourse::where('user_id', $user->id)->get();
 
-                foreach ($courseArr as $course_id) {
-                    $coursePresent = SubscribeCourse::where('user_id', $user->id)->where('course_id', $course_id)->first();
+                    foreach ($courseArr as $course_id) {
+                        $coursePresent = SubscribeCourse::where('user_id', $user->id)->where('course_id', $course_id)->first();
 
-                    if (!$coursePresent) {
-                        $subCourses = new SubscribeCourse();
-                        $subCourses->user_id = $user->id;
-                        $subCourses->course_id = $course_id;
-                        $subCourses->save();
+                        if (!$coursePresent) {
+                            $subCourses = new SubscribeCourse();
+                            $subCourses->user_id = $user->id;
+                            $subCourses->course_id = $course_id;
+                            $subCourses->save();
+                        }
                     }
-                }
-
-                if ($subCourse) {
-                    foreach ($subCourse as $subscribe) {
-                        if (!in_array($subscribe->course_id, $courseArr)) {
-                            SubscribeCourse::find($subscribe->id)->delete();
+                    if ($subCourse) {
+                        foreach ($subCourse as $subscribe) {
+                            if (!in_array($subscribe->course_id, $courseArr)) {
+                                SubscribeCourse::find($subscribe->id)->delete();
+                            }
                         }
                     }
                 }
-
+                    
                 return $this->sendSuccess([], "Student Updated Successfully");
             }
 
             return $this->sendError("Not Found", 404);
         } catch (\Throwable $th) {
-            return $this->sendError("Internal Server Error", 500);
+            // return $this->sendError("Internal Server Error", 500);
+            return $this->sendError($th->getMessage(), 500);
         }
     }
 
@@ -107,19 +115,24 @@ class ProfileController extends BaseController
                 return $this->sendError("Validation Error", 403);
             }
 
-            if ($request->user()->tokenCan('edit-profiles')){
+            if ($request->user()->tokenCan('edit-profiles')) {
                 // store details
                 $user = User::find($request->id);
-                
+
+                // check email already exist
+                if (User::where('email', $request->email)->where('id', '<>', $request->id)->first()) {
+                    return $this->sendError("Email already exists", 409);
+                }
+
                 if ($request->hasFile('avatar')) {
                     $avatar_url = $this->upload('avatar', 'avatar');
                     $user->avatar = $avatar_url;
                 }
-                
+
                 if ($request->role == "teacher") {
                     $user->role_id = 1;
                 }
-                
+
                 $user->name = $request->name;
                 $user->email = $request->email;
                 $user->mobile = $request->mobile;
@@ -128,7 +141,7 @@ class ProfileController extends BaseController
                 $user->country = $request->country;
                 $user->save();
 
-                $courseArr = explode(",",$request->courses);
+                $courseArr = explode(",", $request->courses);
                 $subCourse = SubscribeCourse::where('user_id', $request->id)->get();
 
                 foreach ($courseArr as $course_id) {
@@ -159,11 +172,12 @@ class ProfileController extends BaseController
         }
     }
 
-    public function userProfile(Request $request, $id) {
+    public function userProfile(Request $request, $id)
+    {
         try {
-            if ($request->user()->tokenCan('show-profiles')){
-                $user = User::with('course')->where('id', $id)->first();
-                return $this->sendSuccess($user, "Profile fetched Successfully"); 
+            if ($request->user()->tokenCan('show-profiles')) {
+                $user = User::with('course')->where('is_deleted', 0)->where('id', $id)->first();
+                return $this->sendSuccess($user, "Profile fetched Successfully");
             }
             return $this->sendError("Not Found", 404);
         } catch (\Throwable $th) {
@@ -171,4 +185,26 @@ class ProfileController extends BaseController
         }
     }
 
+    public function userProfileQuiz(Request $request, $id)
+    {
+        try {
+            if ($request->user()->tokenCan('show-profiles')) {
+                $quiz = Quiz::where('user_id', $id)->with('course')->orderBy('updated_at', 'desc')->get();
+                return $this->sendSuccess($quiz, "Profile fetched Successfully");
+            }
+            return $this->sendError("Not Found", 404);
+        } catch (\Throwable $th) {
+            return $this->sendError("Internal Server Error", 500);
+        }
+    }
+
+    public function profileQuiz()
+    {
+        try {
+            $quiz = Quiz::where('user_id', auth()->user()->id)->with('course')->orderBy('updated_at', 'desc')->get();
+            return $this->sendSuccess($quiz, "User Data");
+        } catch (\Throwable $th) {
+            return $this->sendError("Internal Server Error", 500);
+        }
+    }
 }
